+++ src/ocaml_protoc_plugin/field.ml
@@ -7,11 +7,6 @@ type t =
       data : string;
     } (* string, bytes, embedded messages, packed repeated fields *)
   | Fixed_32_bit of Int32.t (* fixed32, sfixed32, float *)
-let show = function
-  | Varint i -> Printf.sprintf "Varint %Ld" i
-  | Fixed_64_bit i -> Printf.sprintf "Fixed_64_bit %Ld" i
-  | Length_delimited { offset; length; data = _} ->  Printf.sprintf "Length_delimited: %d" (length - offset)
-  | Fixed_32_bit i -> Printf.sprintf "Fixed_32_bit %ld" i
 
 let varint v = Varint v
 let fixed_32_bit v = Fixed_32_bit v
@@ -19,3 +14,39 @@ let fixed_64_bit v = Fixed_64_bit v
 let length_delimited ?(offset=0) ?length data =
   let length = Option.value ~default:(String.length data - offset) length in
   Length_delimited {offset; length; data}
+
+
+let pp: Format.formatter -> t -> unit = fun fmt ->
+  function
+  | Varint a0 ->
+    (Format.fprintf fmt "(@[<2>Field.Varint@ ";
+     (Format.fprintf fmt "%LdL") a0;
+     Format.fprintf fmt "@])")
+  | Fixed_64_bit a0 ->
+    (Format.fprintf fmt
+       "(@[<2>Field.Fixed_64_bit@ ";
+     (Format.fprintf fmt "%LdL") a0;
+     Format.fprintf fmt "@])")
+  | Length_delimited
+      { offset = aoffset; length = alength; data = adata } ->
+    (Format.fprintf fmt
+       "@[<2>Field.Length_delimited {@,";
+     (((Format.fprintf fmt "@[%s =@ " "offset";
+        (Format.fprintf fmt "%d") aoffset;
+        Format.fprintf fmt "@]");
+       Format.fprintf fmt ";@ ";
+       Format.fprintf fmt "@[%s =@ " "length";
+       (Format.fprintf fmt "%d") alength;
+       Format.fprintf fmt "@]");
+      Format.fprintf fmt ";@ ";
+      Format.fprintf fmt "@[%s =@ " "data";
+      (Format.fprintf fmt "%S") adata;
+      Format.fprintf fmt "@]");
+     Format.fprintf fmt "@]}")
+  | Fixed_32_bit a0 ->
+    (Format.fprintf fmt
+       "(@[<2>Field.Fixed_32_bit@ ";
+     (Format.fprintf fmt "%ldl") a0;
+     Format.fprintf fmt "@])")
+
+let show : t -> string = Format.asprintf "%a" pp